Moved the keyword to HomepageUtils
