+++ Keywords/homePageUtils/CreateNewBalanceSheetPopup.groovy
@@ -1,4 +1,4 @@
-package commonUtils
+package homePageUtils
 
 import static com.kms.katalon.core.checkpoint.CheckpointFactory.findCheckpoint
 import static com.kms.katalon.core.testcase.TestCaseFactory.findTestCase
